requirements.txt
google-generativeai
pyperclip
